♻️ refactor(test): share the fake Anthropic harness
Move scripted system-test behavior into a plain package with a testing wrapper, and add the message-driven SSE responder for the performance harness.
